refactor: rename BalanceList to AssetList
