+++ src/app/features/asset-list/components/stacks-fungible-token-asset-list.tsx
